1.4.2 重新打包静态库与动态库，解决暂停下载cpu占用过高

+++ KS3SDKIOS/KS3YunSDK/KS3YunSDK/KS3YunSDK/KSYS3/KSYModel/Object/KS3PutObjectRequest.h
@@ -20,7 +20,7 @@
 @property (nonatomic, strong) NSString *expect;
 @property (nonatomic, strong) NSData *data;
 @property (nonatomic, strong) NSInputStream *stream;
-@property (nonatomic, assign, readonly) int32_t expires;
+@property (nonatomic, assign, readonly) int64_t expires;
 @property (nonatomic, strong) NSString *filename;
 @property (nonatomic, strong) NSString *redirectLocation;
 @property (nonatomic, strong) NSString *callbackUrl;
